✨ feat: 코인 구매 내역 및 잔여 코인 조회

+++ src/main/java/com/picktartup/coinservice/entity/CoinTransaction.java
@@ -23,7 +23,7 @@ public class CoinTransaction {
     @Enumerated(EnumType.STRING)
     private TransactionType transactionType;
 
-    private Long coinAmount;
+    private Double coinAmount;
     private LocalDateTime createdAt;
 
     @ManyToOne
